+++ Kata/Models/IProduct.cs
@@ -3,6 +3,6 @@ namespace Kata.Checkout
     public interface IProduct
     {
         char SKU { get; set; }
-        int Price { get; set; }
+        double Price { get; set; }
     }
 }
